+++ user_docs/en/changes.md
@@ -7,7 +7,7 @@
 ### New Features
 
 * New key commands:
-  * Added unassigned commands for vertical and horizontal mouse wheel simulation to enhance navigation on web pages and apps with dynamic content such as Dism++. (#16462, @Cary-Rowen)
+  * Added unassigned commands for vertical and horizontal rotation of the mouse wheel to enhance navigation on web pages and apps with dynamic content, such as Dism++. (#16462, @Cary-Rowen)
 * Error messages referenced with `aria-errormessage` are now reported in Google Chrome and Mozilla Firefox. (#8318)
 * In LibreOffice Writer (version 24.8 and newer), when toggling text formatting (bold, italic, underline, subscript/superscript, alignment) using the corresponding keyboard shortcut, NVDA announces the new formatting attribute (e.g. "Bold on", "Bold off"). (#4248, @michaelweghorn)